+++ out-s-data/pages/index.tsx
@@ -1,6 +1,7 @@
 import Head from 'next/head'
 import Image from 'next/image'
 import styles from '../styles/Home.module.css'
+import CircleVisualization from '../components/circle-virtualization'
 
 export default function Home() {
   return (
@@ -10,6 +11,9 @@ export default function Home() {
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
+      
+      <CircleVisualization views={5} likes={3} title='title1' contents='contents1'/>
+      <CircleVisualization views={10} likes={3} title='title2' contents='contents2'/>
     </div>
   )
 }
